backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
PyYAML==5.4.1
python-dotenv==0.19.0
gunicorn==20.1.0
uvicorn==0.23.0
paho-mqtt==1.6.1 